+++ R/tableby.R
@@ -70,6 +70,9 @@
 #'   \item{
 #'     \code{logrank}: log-rank, the default for time-to-event variables
 #'   }
+#'   \item{
+#'     \code{notest}: no test is performed.
+#'   }
 #' }
 #'
 #' To perform a mixture of asymptotic and rank-based tests on two
@@ -131,7 +134,7 @@ tableby <- function(formula, data, na.action, subset=NULL, weights=NULL, control
     # purposely using na.action instead of temp.call$na.action here
     warning("It appears you're using na.tableby with a one-sided formula... Results may not be what you expect.")
   }
-  special <- c("anova", "kwt", "chisq", "fe", "logrank", "trend")
+  special <- c("anova", "kwt", "chisq", "fe", "logrank", "trend", "notest")
   if (missing(data)) {
     temp.call$formula <- stats::terms(formula, special)
   } else {
@@ -151,7 +154,7 @@ tableby <- function(formula, data, na.action, subset=NULL, weights=NULL, control
   tmp.fun <- function(x, ..., digits = NULL, digits.count = NULL, digits.pct = NULL, cat.simplify = NULL, numeric.simplify = NULL)
   {
     attr(x, "name") <- deparse(substitute(x))
-    attr(x, "stats") <- list(...)
+    attr(x, "stats") <- if(missing(...)) NULL else list(...)
     attr(x, "control.list") <- list(digits = digits, digits.count = digits.count, digits.pct = digits.pct,
                                     cat.simplify = cat.simplify, numeric.simplify = numeric.simplify)
     x
@@ -215,7 +218,7 @@ tableby <- function(formula, data, na.action, subset=NULL, weights=NULL, control
     by.levels <- unique(replace(by.levels, by.levels == "", " "))
   }
 
-  if(length(by.levels) < 2 && attributes(Terms)$response != 0 && control$test)
+  if(length(by.levels) < 2 && attr(Terms, "response") != 0 && control$test)
   {
     warning("The by-variable has fewer than two levels; statistical tests are ignored")
     control$test <- FALSE
@@ -226,13 +229,14 @@ tableby <- function(formula, data, na.action, subset=NULL, weights=NULL, control
     currcol <- modeldf[[eff]]
 
     ## label
-    nameEff <- attributes(currcol)$name
-    if(is.null(nameEff))  nameEff <- names(modeldf)[eff]
-    labelEff <-  attributes(currcol)$label
-    if(is.null(labelEff))  labelEff <- nameEff
+    nameEff <- attr(currcol, "name")
+    if(is.null(nameEff)) nameEff <- names(modeldf)[eff]
+    labelEff <- attr(currcol, "label")
+    if(is.null(labelEff)) labelEff <- nameEff
     statList <- list()
     bystatlist <- list()
     control.list <- attr(currcol, "control.list")
+    attrstats <- attr(currcol, "stats")
 
     ############################################################
 
@@ -279,6 +283,7 @@ tableby <- function(formula, data, na.action, subset=NULL, weights=NULL, control
     } else if(survival::is.Surv(currcol)) {
       ##### Survival (time to event) #######
       xlevels <- NULL
+      if(any(currcol[, 2] %nin% 0:1)) stop("Survival endpoint may not be coded 0/1.")
 
       currstats <- control$surv.stats
       currtest <- control$surv.test
@@ -301,7 +306,7 @@ tableby <- function(formula, data, na.action, subset=NULL, weights=NULL, control
 
     ## if no missings, and control says not to show missings,
     ## remove Nmiss stat fun
-    currstats <- if(length(attributes(currcol)$stats)>0) attributes(currcol)$stats else currstats
+    currstats <- if(is.null(attrstats)) currstats else attrstats
     if(!anyNA(currcol) && "Nmiss" %in% currstats) currstats <- currstats[currstats != "Nmiss"]
     for(statfun in currstats) {
       if(statfun %in% c("countrowpct", "countcellpct", "rowbinomCI"))
@@ -331,7 +336,7 @@ tableby <- function(formula, data, na.action, subset=NULL, weights=NULL, control
 
   if(length(xList) == 0) stop("No x-variables successfully computed.")
 
-  labelBy <- attributes(by.col)$label
+  labelBy <- attr(by.col, "label")
   if(is.null(labelBy)) labelBy <- names(modeldf)[1]
 
   yList <- list()